switch to Int64 in tests

+++ test/project_euler.jl
@@ -78,9 +78,9 @@ using TheAlgorithms.ProjectEuler
     end
 
     @testset "Project Euler: Problem 014" begin
-        @test problem_014(10) == 9
-        @test problem_014(250) == 231
-        @test problem_014(1000000) == 837799
-        @test_throws DomainError problem_014(-1)
+        @test problem_014(Int64(10)) == 9
+        @test problem_014(Int64(250)) == 231
+        @test problem_014(Int64(1000000)) == 837799
+        @test_throws DomainError problem_014(Int64(-1))
     end
 end
